refactor(ping_pong/game): ping_pong game refactoring (#21)
* refactor(ping_pong/game): remove assets.hpp
fix(ping_pong/game/README.md): fix run command

* fix(ping_pong/game/Makefile): fix test directory

* refactor(ping_pong/game): add to screen.* addText and rename add to addMatrix

* fix(ping_pong/game): add compile_commands.json to .gitignore

* fix(ping_pong/game): minifix

* feat(ping_pong/game/README.md): Add gif to README

+++ ping_pong/game/src/gameobj/ball.cpp
@@ -1,10 +1,11 @@
 #include "gameobj/ball.hpp"
-#include "assets.hpp"
 #include <vector>
+#include "screen.hpp"
 
 Ball::Ball(float _speedX, float _speedY, int _width, int _height)
     : width(_width), height(_height), defaultSpeedX(_speedX),
       defaultSpeedY(_speedY) {}
+
 void Ball::tick(struct scrsize ws, size_t frameTimeMs) {
     if (this->posX == -1 || this->posY == -1) {
         this->reset(ws);
@@ -15,8 +16,8 @@ void Ball::tick(struct scrsize ws, size_t frameTimeMs) {
 }
 
 void Ball::reset(struct scrsize ws) {
-    this->posX = ws.width / 2;
-    this->posY = ws.height / 2;
+    this->posX = static_cast<float>(ws.width) / 2;
+    this->posY = static_cast<float>(ws.height) / 2;
 
     this->speedX = this->defaultSpeedX;
     this->speedY = this->defaultSpeedY;
